# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StudyFilter API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /progress:
    get:
      operationId: getProgress
      tags: [progress]
      summary: Get student progress stats
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProgress"

    put:
      operationId: updateProgress
      tags: [progress]
      summary: Update student progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressUpdate"
      responses:
        "200":
          description: Updated progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProgress"
  /progress/activity:
    get:
      operationId: getRecentActivity
      tags: [progress]
      summary: Get recent activity for a session
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recent activity entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"

    post:
      operationId: addActivity
      tags: [progress]
      summary: Add a recent activity entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivityInput"
      responses:
        "201":
          description: Activity recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Activity"
...
  /pomodoro/settings:
    get:
      operationId: getPomodoroSettings
      tags: [pomodoro]
      summary: Get pomodoro settings for a session
      parameters:
        - in: query
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroSettings"
    put:
      operationId: updatePomodoroSettings
      tags: [pomodoro]
      summary: Update pomodoro settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PomodoroSettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroSettings"
  /pomodoro/sessions:
    post:
      operationId: createFocusSession
      tags: [pomodoro]
      summary: Start a focus session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FocusSessionInput"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSessionResponse"
  /pomodoro/sessions/{focusSessionId}:
    put:
      operationId: updateFocusSession
      tags: [pomodoro]
      summary: Update or finish a focus session
      parameters:
        - in: path
          name: focusSessionId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FocusSessionUpdate"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FocusSessionResponse"
  /pomodoro/statistics:
    get:
      operationId: getPomodoroStatistics
      tags: [pomodoro]
      summary: Focus statistics for a session
      parameters:
        - in: query
          name: sessionId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PomodoroStatistics"
components:
  schemas:
...
    StudentProgress:
      type: object
      required: [sessionId, xp, streak, questionsSolved, questionsToday, dailyGoal, accuracy, chaptersCompleted]
      properties:
        sessionId:
          type: string
        xp:
          type: integer
        streak:
          type: integer
        questionsSolved:
          type: integer
        questionsToday:
          type: integer
          description: Questions asked today (resets each day)
        dailyGoal:
          type: integer
        accuracy:
          type: number
        chaptersCompleted:
          type: integer
        lastActiveDate:
          type: ["string", "null"]
    ProgressUpdate:
      type: object
      required: [sessionId]
      properties:
        sessionId:
          type: string
        xpDelta:
          type: integer
        questionsSolvedDelta:
          type: integer
        correctAnswers:
          type: integer
        totalAnswers:
          type: integer
        chapterCompleted:
          type: ["string", "null"]
    ActivityInput:
      type: object
      required: [sessionId, question, subject, classLevel]
      properties:
        sessionId:
          type: string
        question:
          type: string
        subject:
          type: string
        classLevel:
          type: integer
        answerSource:
          type: string
    Activity:
      type: object
      required: [id, sessionId, question, subject, classLevel, createdAt]
      properties:
        id:
          type: integer
        sessionId:
          type: string
        question:
          type: string
        subject:
          type: string
        classLevel:
          type: integer
        answerSource:
          type: ["string", "null"]
        createdAt:
          type: string
...
    PomodoroSettings:
      type: object
      required: [sessionId, focusMinutes, shortBreakMinutes, longBreakMinutes, sessionsBeforeLongBreak, autoStartBreaks, autoStartFocus, dailyGoalMinutes, weeklyGoalMinutes, soundEnabled, soundVolume, musicVolume, appearancePreset]
      properties:
        sessionId:
          type: string
        focusMinutes:
          type: integer
        shortBreakMinutes:
          type: integer
        longBreakMinutes:
          type: integer
        sessionsBeforeLongBreak:
          type: integer
        autoStartBreaks:
          type: boolean
        autoStartFocus:
          type: boolean
        dailyGoalMinutes:
          type: integer
        weeklyGoalMinutes:
          type: integer
        soundEnabled:
          type: boolean
        soundVolume:
          type: number
        musicVolume:
          type: number
        preferredAudioId:
          type: string
          nullable: true
        appearancePreset:
          type: string
    PomodoroSettingsInput:
      type: object
      required: [sessionId]
      properties:
        sessionId:
          type: string
        focusMinutes:
          type: integer
        shortBreakMinutes:
          type: integer
        longBreakMinutes:
          type: integer
        sessionsBeforeLongBreak:
          type: integer
        autoStartBreaks:
          type: boolean
        autoStartFocus:
          type: boolean
        dailyGoalMinutes:
          type: integer
        weeklyGoalMinutes:
          type: integer
        soundEnabled:
          type: boolean
        soundVolume:
          type: number
        musicVolume:
          type: number
        preferredAudioId:
          type: string
          nullable: true
        appearancePreset:
          type: string
    FocusSessionInput:
      type: object
      required: [sessionId, sessionType, plannedSeconds, startedAt]
      properties:
        sessionId:
          type: string
        planId:
          type: integer
          nullable: true
        taskId:
          type: integer
          nullable: true
        sessionType:
          type: string
          enum: [focus, short_break, long_break, custom]
        plannedSeconds:
          type: integer
        subject:
          type: string
          nullable: true
        chapter:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
    FocusSessionUpdate:
      type: object
      required: [sessionId]
      properties:
        sessionId:
          type: string
        actualSeconds:
          type: integer
        status:
          type: string
          enum: [active, paused, completed, skipped, cancelled]
        endedAt:
          type: string
          format: date-time
          nullable: true
    FocusSessionRecord:
      type: object
      required: [id, sessionId, sessionType, plannedSeconds, actualSeconds, startedAt, status]
      properties:
        id:
          type: integer
        sessionId:
          type: string
        planId:
          type: integer
          nullable: true
        taskId:
          type: integer
          nullable: true
        sessionType:
          type: string
        plannedSeconds:
          type: integer
        actualSeconds:
          type: integer
        subject:
          type: string
          nullable: true
        chapter:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
    FocusSessionResponse:
      type: object
      required: [focusSession]
      properties:
        focusSession:
          $ref: "#/components/schemas/FocusSessionRecord"
    SubjectMinutes:
      type: object
      required: [subject, minutes]
      properties:
        subject:
          type: string
        minutes:
          type: integer
    PomodoroStatistics:
      type: object
      required: [todayMinutes, weekMinutes, dailyGoalMinutes, weeklyGoalMinutes, completedSessionsToday, completedSessionsWeek, averageSessionMinutes, currentStreakDays, subjectDistribution]
      properties:
        todayMinutes:
          type: integer
        weekMinutes:
          type: integer
        dailyGoalMinutes:
          type: integer
        weeklyGoalMinutes:
          type: integer
        completedSessionsToday:
          type: integer
        completedSessionsWeek:
          type: integer
        averageSessionMinutes:
          type: integer
        currentStreakDays:
          type: integer
        mostFocusedChapter:
          type: string
          nullable: true
        subjectDistribution:
          type: array
          items:
            $ref: "#/components/schemas/SubjectMinutes"
...